+++ src/main/java/com/hypherionmc/simplerpc/SimpleRPCClient.java
@@ -3,11 +3,16 @@ package com.hypherionmc.simplerpc;
 import com.hypherionmc.craterlib.api.events.client.*;
 import com.hypherionmc.craterlib.core.event.CraterEventBus;
 import com.hypherionmc.craterlib.core.event.annot.CraterEventListener;
+import com.hypherionmc.craterlib.core.platform.ModloaderEnvironment;
 import com.hypherionmc.craterlib.nojang.client.BridgedMinecraft;
 import com.hypherionmc.simplerpc.discord.SimpleRPCCore;
 import com.hypherionmc.simplerpc.enums.GameType;
 import com.hypherionmc.simplerpc.enums.RichPresenceState;
 import com.hypherionmc.simplerpc.util.variables.RPCVariables;
+import shadow.kyori.adventure.text.Component;
+import shadow.kyori.adventure.text.format.NamedTextColor;
+import shadow.kyori.adventure.text.format.Style;
+import shadow.kyori.adventure.text.format.TextDecoration;
 
 /**
  * @author HypherionSA
@@ -15,6 +20,8 @@ import com.hypherionmc.simplerpc.util.variables.RPCVariables;
  * Main Mod Entrypoint for Modloaders
  */
 public final class SimpleRPCClient {
+
+    private static boolean hasShownWarning = false;
 
     public static void setupEvents() {
         CraterEventBus.INSTANCE.registerEventListener(SimpleRPCClient.class);
@@ -36,6 +43,14 @@ public final class SimpleRPCClient {
     @CraterEventListener
     public static void screenOpenEvent(ScreenEvent.Opening event) {
         if (event.getScreen().isTitleScreen()) {
+            if (ModloaderEnvironment.INSTANCE.isModLoaded("craftpresence") && !hasShownWarning) {
+                BridgedMinecraft.getInstance().showWarningScreen(
+                        Component.text("Warning").style(Style.style(NamedTextColor.YELLOW).decorate(TextDecoration.BOLD)),
+                        Component.text("You have both Simple RPC and CraftPresence installed. This will cause your RPC to break and behave badly. Please remove either mod and restart the game")
+                );
+                hasShownWarning = true;
+            }
+
             SimpleRPCCore.INSTANCE.getEvents().setRPCState(RichPresenceState.MAIN_MENU);
         }